Module gg::scene

The Scenes module TODO: Expand

Structs

StackSceneManager
A sample implementation of SceneManager can be used as is for a stack based scene system. The type parameter is the state of the game.

Enums

SceneTransition
Signalling Enum, meant to tell the SceneManager what should happen next.

Traits

Scene
One of the most important traits for a game, the scene is what tells the display what to draw as well as what should happen with the given input.

SceneManager
This trait has to be implemented by the SceneManager that will run your game. A sample implementation is StackSceneManager