Enum gg::scene::SceneTransition

pub enum SceneTransition<T: Sized> {
    Nothing,
    Push(Box<Scene<State=T>>),
    Pop,
    PopUntil(usize),
}

Signalling Enum, meant to tell the SceneManager what should happen next.

Variants

Nothing

Nothing will leave the current Scene on the Stack.

Push

Push will leave the current scene (but not destroy it) and put the new scene on the stack.

Pop

Pop will remove the current Scene from the stack returning to the previous one.

PopUntil

PopUntil will remove scenes until the given scene is found, this is useful to get back to a parent menu for example. This panics if the menu does not exist!