Struct gg::scene::StackSceneManager

pub struct StackSceneManager<T: Sized> {
    // some fields omitted
}

A sample implementation of SceneManager can be used as is for a stack based scene system. The type parameter is the state of the game.